+++ pkg/controllers/componentUpgradeManager.go
@@ -22,16 +22,18 @@ type ComponentUpgradeManager struct {
 	healthManager   healthManager
 	recorder        record.EventRecorder
 	timeout         time.Duration
+	reader          configMapRefReader
 }
 
 // NewComponentUpgradeManager creates a new instance of ComponentUpgradeManager.
-func NewComponentUpgradeManager(componentClient componentInterface, helmClient helmClient, healthManager healthManager, recorder record.EventRecorder, timeout time.Duration) *ComponentUpgradeManager {
+func NewComponentUpgradeManager(componentClient componentInterface, helmClient helmClient, healthManager healthManager, recorder record.EventRecorder, timeout time.Duration, reader configMapRefReader) *ComponentUpgradeManager {
 	return &ComponentUpgradeManager{
 		componentClient: componentClient,
 		helmClient:      helmClient,
 		healthManager:   healthManager,
 		recorder:        recorder,
 		timeout:         timeout,
+		reader:          reader,
 	}
 }
 
@@ -60,6 +62,7 @@ func (cum *ComponentUpgradeManager) Upgrade(ctx context.Context, component *k8sv
 		HelmClient:     cum.helmClient,
 		Timeout:        cum.timeout,
 		YamlSerializer: yaml.NewSerializer(),
+		Reader:         cum.reader,
 	})
 	if err != nil {
 		return fmt.Errorf("failed to get helm chart spec: %w", err)