+++ pkg/controllers/requeableErrors_test.go
@@ -2,10 +2,11 @@ package controllers
 
 import (
 	"errors"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func Test_genericRequeueableError_Unwrap(t *testing.T) {
@@ -33,24 +34,23 @@ func Test_genericRequeueableError_Error(t *testing.T) {
 
 func Test_genericRequeueableError_GetRequeueTime(t *testing.T) {
 	type args struct {
-		requeueTime time.Duration
+		requeueTime        time.Duration
+		defaultRequeueTime time.Duration
 	}
 	tests := []struct {
 		name string
 		args args
 		want time.Duration
 	}{
-		// double the value until the threshold jumps in
-		{"1st interval", args{0 * time.Second}, 15 * time.Second},
-		{"2nd interval", args{15 * time.Second}, 30 * time.Second},
-		{"3rd interval", args{30 * time.Second}, 1 * time.Minute},
-		{"11th interval", args{128 * time.Minute}, 256 * time.Minute},
-		{"cutoff interval ", args{256 * time.Minute}, 6 * time.Hour},
+		{"always return defaultRequeueTime case 1", args{0 * time.Second, 5 * time.Second}, 5 * time.Second},
+		{"always return defaultRequeueTime case 2", args{15 * time.Second, 5 * time.Second}, 5 * time.Second},
+		{"always return defaultRequeueTime case 3", args{30 * time.Second, 5 * time.Second}, 5 * time.Second},
+		{"always return defaultRequeueTime case 3", args{256 * time.Second, 5 * time.Second}, 5 * time.Second},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			due := &genericRequeueableError{}
-			assert.Equalf(t, tt.want, due.GetRequeueTime(tt.args.requeueTime), "getRequeueTime(%v)", tt.args.requeueTime)
+			assert.Equalf(t, tt.want, due.GetRequeueTime(tt.args.requeueTime, tt.args.defaultRequeueTime), "getRequeueTime(%v)", tt.args.requeueTime)
 		})
 	}
 }