+++ pkg/helm/componentChart.go
@@ -6,7 +6,6 @@ import (
 	"time"
 
 	componentV1 "github.com/cloudogu/k8s-component-lib/api/v1"
-
 	"github.com/cloudogu/k8s-component-operator/pkg/helm/client"
 	"github.com/cloudogu/k8s-component-operator/pkg/helm/client/values"
 	"github.com/cloudogu/k8s-component-operator/pkg/labels"
@@ -26,6 +25,7 @@ type HelmChartCreationOpts struct {
 	HelmClient     ChartGetter
 	Timeout        time.Duration
 	YamlSerializer yaml.Serializer
+	Reader         configMapRefReader
 }
 
 type Mapping struct {
@@ -57,18 +57,20 @@ func GetHelmChartSpec(ctx context.Context, c *componentV1.Component, opts ...Hel
 	timeout := defaultHelmClientTimeoutMins
 	var chartGetter ChartGetter
 	var yamlSerializer yaml.Serializer
+	var reader configMapRefReader
 	if len(opts) > 0 {
 		timeout = opts[0].Timeout
 		chartGetter = opts[0].HelmClient
 		yamlSerializer = opts[0].YamlSerializer
+		reader = opts[0].Reader
 	}
 
 	chartSpec := &client.ChartSpec{
-		ReleaseName: c.Spec.Name,
-		ChartName:   GetHelmChartName(c),
-		Namespace:   deployNamespace,
-		Version:     c.Spec.Version,
-		ValuesYaml:  c.Spec.ValuesYamlOverwrite,
+		ReleaseName:         c.Spec.Name,
+		ChartName:           GetHelmChartName(c),
+		Namespace:           deployNamespace,
+		Version:             c.Spec.Version,
+		ValuesYamlOverwrite: c.Spec.ValuesYamlOverwrite,
 		// Rollback to previous release on failure.
 		Atomic: true,
 		// This timeout prevents context exceeded errors from the used k8s client from the helm library.
@@ -88,6 +90,11 @@ func GetHelmChartSpec(ctx context.Context, c *componentV1.Component, opts ...Hel
 		chartSpec.MappedValuesYaml, err = getMappedValuesYaml(ctx, c, chartSpec, chartGetter, yamlSerializer)
 		if err != nil {
 			return nil, fmt.Errorf("failed to create mapped values: %w", err)
+		}
+
+		chartSpec.ValuesConfigRefYaml, err = reader.GetValues(ctx, c.Spec.ValuesConfigRef)
+		if err != nil {
+			return nil, fmt.Errorf("failed to create values config references: %w", err)
 		}
 	}